+++ Backend/src/main/java/pl/covidSimulation/entity/SimulationData.java
@@ -20,7 +20,7 @@ public class SimulationData {
     private Integer p;
     private Integer i; // initial number of infected
     private Integer r; //An indicator of how many people one infected person infects,
-    private Integer m; // mortality rate
+    private Double m; // mortality rate
     private Integer ti; // The number of days that elapse between infection and recovery of the patient
     private Integer tm; //The number of days that elapse between infection and the death of the patient
     private Integer ts; // Number of days for which the simulation is to be carried out
